Better lighting

+++ vinox-common/src/collision/aabb.rs
@@ -43,11 +43,8 @@ pub fn aabb_vs_world(
                 );
                 if let Some(chunk_entity) = current_chunks.get_entity(ChunkPos(check_chunk_pos)) {
                     if let Ok(chunk) = chunks.get(chunk_entity) {
-                        let block_data: BlockData = chunk.get(
-                            check_block_cpos.x as usize,
-                            check_block_cpos.y as usize,
-                            check_block_cpos.z as usize,
-                        );
+                        let block_data: BlockData =
+                            chunk.get(check_block_cpos.x, check_block_cpos.y, check_block_cpos.z);
                         let voxel_pos = voxel_to_global_voxel(check_block_cpos, check_chunk_pos);
                         if !block_data.is_empty(block_table) {
                             let block_aabb = Aabb {
@@ -174,11 +171,7 @@ pub fn aabb_vs_world(
         let (chunk_pos, voxel_cpos) = world_to_voxel(blocked_pos);
         if let Some(chunk_entity) = current_chunks.get_entity(ChunkPos(chunk_pos)) {
             if let Ok(chunk) = chunks.get(chunk_entity) {
-                let block_data: BlockData = chunk.get(
-                    voxel_cpos.x as usize,
-                    voxel_cpos.y as usize,
-                    voxel_cpos.z as usize,
-                );
+                let block_data: BlockData = chunk.get(voxel_cpos.x, voxel_cpos.y, voxel_cpos.z);
                 return block_data.is_empty(block_table);
             }
         }
